+++ test/Recipe-test.js
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 
 import Recipe from '../src/classes/Recipe';
-import { recipeData } from '../src/data/recipes';
-import { ingredientsData } from '../src/data/ingredients';
+import { recipeData } from '../src/testData/recipes';
+import { ingredientsData } from '../src/testData/ingredients';
 
 let recipe;
